新页面

+++ app/src/test/java/com/example/healthy/ExampleUnitTest.kt
@@ -23,7 +23,7 @@ class ExampleUnitTest {
     fun testHeartData() {
         val heart = HeartOneData()
         for (i in heart.bodyData.indices) {
-            heart.bodyData[i] = i.toShort()
+            heart.bodyData[i] = i.toInt()
         }
         val result = heart.getData()
         assertEquals(result[0].size, 5)
@@ -33,7 +33,7 @@ class ExampleUnitTest {
     fun testHeartThree() {
         val three = HeartThreeData()
         for (i in three.bodyData.indices) {
-            three.bodyData[i] = i.toShort()
+            three.bodyData[i] = i
         }
         val result = three.getData()
         assertEquals(result[0].size, 2)
